Delete Duplicates/Elson-main/Elson/backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.19
pydantic==2.1.1
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
pytest==7.4.0
python-dotenv==1.0.0
websockets==11.0.3
pandas==2.0.3
numpy==1.25.1
scikit-learn==1.3.0
torch==2.0.1
yfinance==0.2.28
aiohttp==3.8.5
email-validator==2.0.0
alembic==1.11.1
redis==4.6.0
aioredis==2.0.1

# Security-related packages
cryptography==41.0.3
bcrypt==4.0.1
python-multipart==0.0.6
pyjwt==2.8.0
fastapi-limiter==0.1.5
slowapi==0.1.8

# Monitoring and observability
prometheus-client==0.17.1
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
sentry-sdk==1.33.1

# Testing
pytest-asyncio==0.21.1
httpx==0.24.1
faker==19.3.0

# Quantum model integration
qiskit==1.4.2
qiskit-aer==0.17.0
qiskit-machine-learning==0.8.2
